backend/requirements.txt
azure-functions
azure-storage-blob
pandas
matplotlib
seaborn
numpy